+++ src/main/java/com/telepathicgrunt/repurposedstructures/mixin/MinecraftServerMixin.java
@@ -1,5 +1,6 @@
 package com.telepathicgrunt.repurposedstructures.mixin;
 
+import com.google.common.collect.ImmutableMap;
 import com.mojang.authlib.GameProfileRepository;
 import com.mojang.authlib.minecraft.MinecraftSessionService;
 import com.mojang.datafixers.DataFixer;
@@ -13,9 +14,11 @@ import net.minecraft.util.registry.DynamicRegistries;
 import net.minecraft.util.registry.Registry;
 import net.minecraft.world.biome.Biome;
 import net.minecraft.world.chunk.listener.IChunkStatusListenerFactory;
+import net.minecraft.world.gen.DimensionSettings;
 import net.minecraft.world.gen.feature.ConfiguredFeature;
 import net.minecraft.world.storage.IServerConfiguration;
 import net.minecraft.world.storage.SaveFormat;
+import org.apache.logging.log4j.Level;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -77,7 +80,24 @@ public class MinecraftServerMixin {
                 RepurposedStructures.addFeaturesAndStructuresToBiomes(registryManager.getOptional(Registry.BIOME_KEY).get(), biome, Objects.requireNonNull(registryManager.getOptional(Registry.BIOME_KEY).get().getKey(biome)), allBiomeBlacklists);
 
                 //add our structure spacing to all chunkgenerators
-                registryManager.get(Registry.NOISE_SETTINGS_WORLDGEN).forEach(registryManager -> registryManager.getStructuresConfig().structures.putAll(RSFeatures.RS_STRUCTURES));
+                for(DimensionSettings dimensionSettings : registryManager.get(Registry.NOISE_SETTINGS_WORLDGEN)){
+                    if(dimensionSettings.getStructuresConfig().structures instanceof ImmutableMap){
+                        //A mod structures spacing is using an immutable map. Time to shame that mod lol
+                        //By shame I mean, print to the logs so I can make a bug report to that mod to not immutable map this.
+                        RepurposedStructures.LOGGER.log(Level.WARN, "--------------------------------------------------------------------");
+                        RepurposedStructures.LOGGER.log(Level.WARN, "");
+                        RepurposedStructures.LOGGER.log(Level.WARN, "Error: "+registryManager.get(Registry.NOISE_SETTINGS_WORLDGEN).getKey(dimensionSettings)+" 's " +
+                                                                            "Map<Structure<?>, StructureSeparationSettings> structures was unabled to be modified by Repurposed Structures " +
+                                                                            "because the map is an immutable map. Please let either mod author know this so that the map can be changed to be " +
+                                                                            "a normal mutable map. Otherwise, Repurposed Structures stuff will spawn in every single chunk of that mod's dimension " +
+                                                                            "as the map is used to store structure's spacing configs.");
+                        RepurposedStructures.LOGGER.log(Level.WARN, "");
+                        RepurposedStructures.LOGGER.log(Level.WARN, "--------------------------------------------------------------------");
+                    }
+                    else{
+                        dimensionSettings.getStructuresConfig().structures.putAll(RSFeatures.RS_STRUCTURES);
+                    }
+                }
             }
         }
     }